refactor: update Twig navigation to allow parameters
Navigation was still using the deprecated buildNavigation method.
In addition, it was impossible to configure custom parameters, which is
possible when you use the Pimcore build_nav Twig function.

The logic was refactored to work similar to the Pimcore build_nav. The old
deprecated setup still works, but parameters can (and should) be used now.
In addition, when configuring the pageCallback parameter, it is now called
within the members callback, which also forwards the restriction element,
allowing you to override the restriction settings and apply other logic.